feat: 할인 도메인 구성

+++ src/main/java/org/shoestore/promotion/model/PromotionRule.java
@@ -1,5 +1,5 @@
 package org.shoestore.promotion.model;
 
 public interface PromotionRule {
-    boolean validateRule(double price, int discountValue);
+    boolean validateRule(double price, double discountValue);
 }
